refactor: move .test.js files to /tests
