+++ src/tests/utils/helpers/ui.helpers.test.js
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react'
-import { getLinkAsButton } from './ui.helpers'
+import { getLinkAsButton } from '../../../utils/helpers/ui.helpers'
 
 describe('uiUtils', () => {
   describe('getLinkAsButton', () => {
